pyproject.toml
[project]
name = "ghost"
version = "0.1.0"
description = "Open-source Ghost continuity system for LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name="Ghost"}]

dependencies = [
  "fastapi>=0.111,<0.115",
  "uvicorn>=0.30,<0.31",
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "typer>=0.12,<0.13",
  "httpx>=0.27,<0.28",
  "requests>=2.32,<3.0",
  "beautifulsoup4>=4.12,<5.0",
  "urllib3>=2.2,<3.0",
  "python-dotenv>=1.0,<2.0",
  "prometheus-client>=0.20,<0.21",
  "orjson>=3.10,<4.0",
  "numpy>=1.26,<2.0",
  "scipy>=1.12,<2.0",
  "structlog>=24.1,<25.0",
  "jinja2>=3.1,<4.0",
  "starlette>=0.37,<0.39",
  "aiofiles>=23.2,<24.0",
  "PyNaCl>=1.5,<2.0",
]

[project.optional-dependencies]
prod = [
  "ruff>=0.5,<0.6",
  "mypy>=1.10,<2.0",
]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<0.24",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5,<0.6",
  "mypy>=1.10,<2.0",
]
faiss = ["faiss-cpu>=1.8,<2.0"]
playwright = ["playwright>=1.47,<2.0"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app", "cli.py", "tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_any_generics = true
exclude = ["app/ui/.*"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["app"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
